Move validateUrl to widget utils
